LPS-125549 SF, rename package
