+++ modules/apps/petra/petra-log4j/src/main/java/com/liferay/petra/log4j/internal/Log4JConfigurationUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.log4j.internal.configuration;
+package com.liferay.petra.log4j.internal;
 
 import com.liferay.portal.kernel.io.unsync.UnsyncStringReader;
