+++ modules/apps/petra/petra-log4j/src/test/java/com/liferay/petra/log4j/internal/Log4JConfigurationUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.log4j.internal.configuration;
+package com.liferay.petra.log4j.internal;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.test.rule.CodeCoverageAssertor;
